Hide the "Gateway" sidebar link
This commit hides the "Gateway" sidebar link in the dashboard if the
`linkerd-extension` is not installed. If a user happens to navigate to
the Gateway page anyway, we display a CTA (Call to Action) that informs
the user that they would need to run the multicluster install command.

This change includes a new endpoint in the dashboard server; `GET
/api/extensions`. This endpoint returns a list of all currently
installed `linkerd.io/extension` and the corresponding namespace each
extension is installed in. The dashboard uses this endpoint to detect
whether it needs to hide the navigation link and whether to displaye the
CTA.

Fixes #5330

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>